Functions
Deleted the struct. Added some constant functions that define the left child, right child and the parent. After this I developed the build heap function and the max heap function. There is a main that has shell code for reading in the rest of the file. The remaining operations must still be developed.